Added log out button, changed dimensions, moved files outside frameworks package.
