+++ src/main/java/Interface_and_Adapters/TabPanel.java
@@ -25,9 +25,9 @@ public class TabPanel extends JPanel{
         JTabbedPane tabs = new JTabbedPane();
 
         RestaurantDataAccess res;
-        res = new RestaurantFileReader("src/main/java/Frameworks_and_Drivers/Restaurant.csv");
+        res = new RestaurantFileReader("./Restaurant.csv");
 
-        DishDataAccess dish = new DishFileReader("src/main/java/Frameworks_and_Drivers/Dishes.csv");
+        DishDataAccess dish = new DishFileReader("./Dishes.csv");
 
 
         RestaurantPresenter presenter = new RestaurantFormatted();
